Taxonomy based cleaning

Taxonomy based cleaning

tax_no_epithet(x, name = NULL, drop = TRUE)

Arguments

x
(data.frame) A data.frame1

name
(character) Taxonomic name field Optional. See Details.

drop
(logical) Drop bad data points or not. Either way, we parse out bade data points as an attribute you can access. Default: TRUE

Value

Returns a data.frame, with attributes

Examples

if (requireNamespace("rgbif", quietly = TRUE) && interactive()) {
   library("rgbif")
   res <- rgbif::occ_data(limit = 200)$data
} else {
   res <- sample_data_3
}

# Remove records where names don't have genus + epithet
## so removes those with only genus and those with no name (NA or NULL)
NROW(res)

#> [1] 200

df <- dframe(res) %>% tax_no_epithet(name = "name")
NROW(df)

#> [1] 190

attr(df, "name_var")

#> [1] "name"

attr(df, "tax_no_epithet")

#> # A tibble: 10 x 5
#>    name  decimalLatitude decimalLongitude        key eventDate                  
#>    <chr>           <dbl>            <dbl>      <int> <chr>                      
#>  1 NA               48.7            2.28  1228053209 2016-01-08T23:00:00.000+00…
#>  2 NA               36.2           -6.07  1229956632 2016-01-18T23:00:00.000+00…
#>  3 NA               43.0            0.410 1229959363 2016-01-23T23:00:00.000+00…
#>  4 NA               38.7         -123.    1233599876 2016-01-22T23:00:00.000+00…
#>  5 NA              -44.7          171.    1234563183 2016-01-22T23:00:00.000+00…
#>  6 NA              -36.9          175.    1234563247 2016-01-08T23:00:00.000+00…
#>  7 NA              -36.9          175.    1234563254 2016-01-08T23:00:00.000+00…
#>  8 NA              -37.1          175.    1234563259 2016-01-09T23:00:00.000+00…
#>  9 NA              -36.9          175.    1234563264 2016-01-08T23:00:00.000+00…
#> 10 NA              -39.3          174.    1234563300 2016-01-10T23:00:00.000+00…