Taxonomy based cleaning

Taxonomy based cleaning

tax_no_epithet(x, name = NULL, drop = TRUE)

Arguments

x
(data.frame) A data.frame

name
(character) Taxonomic name field Optional. See Details.

drop
(logical) Drop bad data points or not. Either way, we parse out bade data points as an attribute you can access. Default: TRUE

Value

Returns a data.frame, with attributes

Examples

if (requireNamespace("rgbif", quietly = TRUE)) {
   library("rgbif")
   res <- rgbif::occ_data(limit = 200)$data
} else {
   res <- sample_data_3
}

# Remove records where names don't have genus + epithet
## so removes those with only genus and those with no name (NA or NULL)
NROW(res)

#> [1] 200

df <- dframe(res) %>% tax_no_epithet(name = "name")
NROW(df)

#> [1] 190

attr(df, "name_var")

#> [1] "name"

attr(df, "tax_no_epithet")

#> # A tibble: 10 x 106
#>    key   scientificName decimalLatitude decimalLongitude issues datasetKey
#>    <chr> <chr>                    <dbl>            <dbl> <chr>  <chr>     
#>  1 1989… Lepidoptera               25.0             122. txmat… e0b8cb67-…
#>  2 1989… Geometridae               24.0             121. txmat… e0b8cb67-…
#>  3 1989… Lepidoptera               24.2             121. txmat… e0b8cb67-…
#>  4 1989… Psychidae                 23.8             121. txmat… e0b8cb67-…
#>  5 1989… Lepidoptera               24.8             121. txmat… e0b8cb67-…
#>  6 1989… Lepidoptera               22.9             121. txmat… e0b8cb67-…
#>  7 1989… Pyralidae                 24.4             121. txmat… e0b8cb67-…
#>  8 1989… Lepidoptera               22.9             121. txmat… e0b8cb67-…
#>  9 1989… Lepidoptera               25.0             122. txmat… e0b8cb67-…
#> 10 1989… Erebidae                  22.9             121. txmat… e0b8cb67-…
#> # … with 100 more variables: publishingOrgKey <chr>, networkKeys <list>,
#> #   installationKey <chr>, publishingCountry <chr>, protocol <chr>,
#> #   lastCrawled <chr>, lastParsed <chr>, crawlId <int>, basisOfRecord <chr>,
#> #   individualCount <int>, lifeStage <chr>, taxonKey <int>, kingdomKey <int>,
#> #   phylumKey <int>, classKey <int>, orderKey <int>, familyKey <int>,
#> #   genusKey <int>, speciesKey <int>, acceptedTaxonKey <int>,
#> #   acceptedScientificName <chr>, kingdom <chr>, phylum <chr>, order <chr>,
#> #   family <chr>, genus <chr>, species <chr>, genericName <chr>,
#> #   specificEpithet <chr>, taxonRank <chr>, taxonomicStatus <chr>, year <int>,
#> #   month <int>, day <int>, eventDate <chr>, lastInterpreted <chr>,
#> #   license <chr>, geodeticDatum <chr>, class <chr>, countryCode <chr>,
#> #   country <chr>, eventID <chr>, identifier <chr>, sampleSizeUnit <chr>,
#> #   associatedReferences <chr>, county <chr>, municipality <chr>,
#> #   locality <chr>, gbifID <chr>, occurrenceID <chr>, vernacularName <chr>,
#> #   institutionCode <chr>, samplingProtocol <chr>, verbatimElevation <chr>,
#> #   identifiedBy <chr>, collectionID <chr>, sampleSizeValue <chr>,
#> #   infraspecificEpithet <chr>, continent <chr>, stateProvince <chr>,
#> #   recordNumber <chr>, verbatimEventDate <chr>, datasetName <chr>,
#> #   georeferenceRemarks <chr>, verbatimLocality <chr>, collectionCode <chr>,
#> #   catalogNumber <chr>, recordedBy <chr>, occurrenceRemarks <chr>,
#> #   institutionID <chr>, samplingEffort <chr>, eventRemarks <chr>,
#> #   locationAccordingTo <chr>, locationRemarks <chr>, fieldNotes <chr>,
#> #   ownerInstitutionCode <chr>, identificationRemarks <chr>,
#> #   coordinateUncertaintyInMeters <dbl>, modified <chr>,
#> #   dynamicProperties <chr>, occurrenceStatus <chr>, locationID <chr>,
#> #   behavior <chr>, rightsHolder <chr>, language <chr>, type <chr>,
#> #   coordinatePrecision <dbl>, organismQuantity <chr>,
#> #   identificationVerificationStatus <chr>, taxonID <chr>,
#> #   organismQuantityType <chr>, acceptedNameUsageID <chr>,
#> #   identificationID <chr>, references <chr>,
#> #   `http://unknown.org/occurrenceDetails` <chr>, rights <chr>,
#> #   eventTime <chr>, dateIdentified <chr>, informationWithheld <chr>,
#> #   name <chr>